+++ boot/src/main/java/com/spring/gameloft/service/AlbumService.java
@@ -8,13 +8,13 @@ public interface AlbumService {
 
     List<Album> getAllAlbums(Long singerId);
 
-    Album getAlbum(Long singerId, Long id);
+    Album getAlbum(Long singerId, Integer id);
 
     Album create(Long singerId, Album album);
 
-    Album update(Long singerId, Long id, Album album);
+    Album update(Long singerId, Integer id, Album album);
 
     List<Album> getAlbumsByTitle(Long singerId, String title);
 
-    void delete(Long singerId, Long id);
+    void delete(Long singerId, Integer id);
 }
